fix crashing bug
"Comparison method violates its general contract!"

Misunderstood the docs on compareToIgnoreCase, thought only values that came out were 10, 0, -10, but actually it could be lots of different numbers (some less than +-5). So sometimes location_offset would overpower the alphabetic_offset when it wasn't supposed to causing a "violation in the general contract".

Didn't cause a problem for a long time, but when I was in cowfield in lumbridge, lots of mobdrops from the cows and a nearby cabbage field would trigger this crash.